+++ PROJECTE DE CREACIÓ ALEX XENIA ERNEST/Assets/_Scripts/IDamageTaker.cs
@@ -1,4 +1,4 @@
 ﻿public interface IDamageTaker
 {
-    void TakeDamage(float amount);
+    void TakeDamage(int amount);
 }
